+++ Assets/EVE/Scripts/Menu/Buttons/SessionParametersButtons.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using Assets.EVE.Scripts.Utils;
 using UnityEngine;
 using UnityEngine.UI;
 
@@ -24,18 +25,34 @@ namespace Assets.EVE.Scripts.Menu.Buttons
 
         public void DisplaySessionParameters()
         {
+            Utils.MenuUtils.ClearList(_dynamicField);
+
             var experimentParameters = _menuManager.GetExperimentParameterList();
 
 
             foreach (var experimentParameter in experimentParameters)
             {
-                var gObject = Instantiate(Resources.Load("Prefabs/Menus/TextAndFieldNoXButton")) as GameObject;
-                Utils.PlaceElement(gObject, _dynamicField);
+                var gObject = GameObjectUtils.InstatiatePrefab("Prefabs/Menus/TextAndFieldNoXButton");
+                gObject.GetComponentInChildren<InputField>().onEndEdit.AddListener((string text) =>
+                {
+                    StoreSessionParameter(experimentParameter,text);
+                });
+                Utils.MenuUtils.PlaceElement(gObject, _dynamicField);
                 gObject.transform.Find("FieldName").GetComponent<Text>().text = experimentParameter;
                 if (_launchManager.SessionParameters.ContainsKey(experimentParameter))
                     gObject.transform.Find("InputField").Find("Placeholder").GetComponent<Text>().text =
                         _launchManager.SessionParameters[experimentParameter];
             }
+        }
+
+        public void StoreSessionParameter(string sessionParameter,string value)
+        {
+            if (value.Equals(""))
+            {
+                Debug.LogWarning("Empty input for session parameter " + sessionParameter);
+                return;
+            }
+            _launchManager.changeSessionsParameter(sessionParameter, value);
         }
     }
 }