+++ CHANGE_LOG.md
@@ -8,4 +8,17 @@ Find an overview of all new additions.
 
 New features for the energyscape experiments
 
--New Question Type: VisualStimuli
+- New Question Type: VisualStimuli
+
+###Menu Overhaul
+
+The menu is standardised to make expansion and use more simple
+
+- The loader is renamed launcher (to match with the LaunchManager)
+- All menus are now moved to the launcher
+  - The evaluation scene is deleted and contents moved to the launcher
+  - The questionnaire scene is deleted and contents are moved to the launcher
+- All menus are instantiated instead of kept in the background
+- All function used within one menu panel are now on the Menu Game Object
+
+